fix: repair broken relative imports in moved src test specs
#410/#411/#412 moved src/**/*.spec.ts into test/src/** but merged with their
relative imports/mocks still pointing at the pre-move locations — and with the
`lint-and-build` check (which runs `yarn test`) red. Codex had flagged the exact
lines as CHANGES REQUESTED on each PR. Result: 51 spec files failed module
resolution on main (e.g. test/src/router/index.spec.ts importing "./index").

Each import was re-resolved against the spec's original src/ location and rewritten
relative to its new test/src/ location, so cross-directory imports (a component test
mocking a composable) land correctly too, not just same-dir ones. Extension style is
preserved (.js/.vue kept as authored). No src/ or server/ files touched — test-only.

Restores the full suite to green (123 files, 1265 tests).
