+++ test/src/components/LauncherCell.spec.ts
@@ -1,10 +1,10 @@
 import { describe, it, expect, vi } from "vitest";
 import { mount } from "@vue/test-utils";
-import LauncherCell from "../../src/components/LauncherCell.vue";
+import LauncherCell from "../../../src/components/LauncherCell.vue";
 
 // Stub the terminal so no xterm/WebSocket is needed; it just forwards the props the
 // cell passes and can emit session/exit.
-vi.mock("./Terminal.vue", () => ({
+vi.mock("../../../src/components/Terminal.vue", () => ({
   default: {
     name: "TerminalView",
     props: ["persistKey", "sessionId", "connectKey", "cwd", "launcher"],
