+++ test/src/components/PrsOverlay.spec.ts
@@ -1,10 +1,10 @@
 import { describe, it, expect, vi } from "vitest";
 import { ref } from "vue";
 import { mount, flushPromises } from "@vue/test-utils";
-import PrsOverlay from "../../src/components/PrsOverlay.vue";
+import PrsOverlay from "../../../src/components/PrsOverlay.vue";
 
 // The view is route-driven; stub usePrsView so the overlay is "open" without a router.
-vi.mock("../../src/composables/usePrsView", () => ({
+vi.mock("../../../src/composables/usePrsView", () => ({
   usePrsView: () => ({ isOpen: ref(true), close: vi.fn() }),
 }));
