+++ test/src/components/TerminalCell.spec.ts
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { mount, flushPromises } from "@vue/test-utils";
 import { nextTick } from "vue";
-import TerminalCell from "../../src/components/TerminalCell.vue";
+import TerminalCell from "../../../src/components/TerminalCell.vue";
 
 // Capture the "sessions" pub/sub callback so tests can push activity directly.
 let captured: ((data: unknown) => void) | null = null;
-vi.mock("../../src/composables/usePubSub", () => ({
+vi.mock("../../../src/composables/usePubSub", () => ({
   usePubSub: () => ({
     subscribe: (_channel: string, cb: (data: unknown) => void) => {
       captured = cb;
@@ -16,7 +16,7 @@ vi.mock("../../src/composables/usePubSub", () => ({
 
 // Stub the terminal so no xterm/WebSocket is needed; expose terminate() since
 // the cell's close() calls it.
-vi.mock("./Terminal.vue", () => ({
+vi.mock("../../../src/components/Terminal.vue", () => ({
   default: {
     name: "TerminalView",
     props: ["sessionId", "connectKey", "cwd", "hideHeader"],
