+++ test/src/components/terminalViewActive.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { terminalManagesAttention, terminalViewActive } from "../../src/components/terminalViewActive.js";
+import { terminalManagesAttention, terminalViewActive } from "../../../src/components/terminalViewActive.js";
 
 describe("terminalManagesAttention", () => {
   it("a Claude/Codex session terminal manages attention", () => {
