+++ test/src/components/wsUrl.spec.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { buildTerminalWsUrl, buildRunWsUrl, buildLaunchWsUrl, buildCodexWsUrl } from "../../src/components/wsUrl.js";
+import { buildTerminalWsUrl, buildRunWsUrl, buildLaunchWsUrl, buildCodexWsUrl } from "../../../src/components/wsUrl.js";
 
 describe("buildTerminalWsUrl", () => {
   it("single view: session only, no gui=0", () => {
