+++ test/src/composables/useTerminalConnections.spec.ts
@@ -74,7 +74,7 @@ class FakeWebSocket {
   }
 }
 
-import * as conn from "./useTerminalConnections";
+import * as conn from "../../../src/composables/useTerminalConnections";
 
 const target = (sessionId: string | null) => ({ sessionId, cwd: "/typed", devTerminal: false, command: null, launcher: null });
